Replace open-chrome-app with open-web-app
open-web-app uses a new "Browser Utilities" library that works with
both Google Chrome and Safari and dynamically determines the default
browser from lauch control settings, or from Choosy config, if Choosy
is the default browser.

This library (and open-web-app) is more general than needed for karabiner
config alone, as it is being reused in Keyboard Maestro macros, Typinator
expansions and other bash scripts.